requirements.txt
pygame==2.5.2
requests==2.31.0
Pillow==10.0.0
openai
python-dotenv


